requirements.txt
spacy>=3.7.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pdfplumber>=0.10.0
python-docx>=0.8.11
matplotlib>=3.7.0
seaborn>=0.12.0
